# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.ChangePasswordRequest:
    properties:
      confirm_new_password:
        type: string
      current_password:
        type: string
      new_password:
        example: Senha@123
        minLength: 8
        type: string
    required:
    - confirm_new_password
    - current_password
    - new_password
    type: object
  auth.DeactivateRequest:
    properties:
      current_password:
        type: string
    required:
    - current_password
    type: object
  auth.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      refresh_token:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/auth.UserResponse'
    type: object
  auth.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.RegisterRequest:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      password:
        example: Senha@123
        minLength: 8
        type: string
    required:
    - confirm_password
    - email
    - name
    - password
    type: object
  auth.ResetPasswordRequest:
    properties:
      confirm_password:
        type: string
      new_password:
        example: Senha@123
        minLength: 8
        type: string
      token:
        type: string
    required:
    - confirm_password
    - new_password
    - token
    type: object
  auth.StatusUpdateRequest:
    properties:
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - status
    type: object
  auth.UserResponse:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  response.FieldError:
    properties:
      error:
        type: string
      field:
        type: string
    type: object
  response.Standard:
    properties:
      data: {}
      errors:
        items:
          $ref: '#/definitions/response.FieldError'
        type: array
      message:
        type: string
      meta: {}
      status:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: Este serviço é o Provedor Central de Identidade (IAM)...
  title: Auth API Microservice (Chameleon System)
  version: "1.0"
paths:
  /admin/users/{id}/status:
    put:
      consumes:
      - application/json
      description: Permite ao Admin banir, suspender ou reativar um usuário.
      parameters:
      - description: ID do usuário a ser atualizado
        in: path
        name: id
        required: true
        type: string
      - description: Novo status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.StatusUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Standard'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Standard'
      security:
      - ApiKeyAuth: []
      summary: Atualiza o status de um usuário (Admin-only)
      tags:
      - Admin
  /auth/change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados para alteração de senha
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
      security:
      - ApiKeyAuth: []
      summary: Altera a senha do usuário logado
      tags:
      - Auth
  /auth/deactivate:
    post:
      consumes:
      - application/json
      description: Exige a senha atual para confirmar a intenção e desativa o status
        do usuário (soft delete).
      parameters:
      - description: Senha atual do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.DeactivateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Standard'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Standard'
      security:
      - ApiKeyAuth: []
      summary: Desativa a própria conta do usuário
      tags:
      - Auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Recebe o e-mail do usuário e envia um token de reset
      parameters:
      - description: E-mail do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Iniciar recuperação de senha
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credenciais do usuário para login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Standard'
            - properties:
                data:
                  $ref: '#/definitions/auth.LoginResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Autenticar usuário
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Adiciona o JWT à blacklist, invalidando a sessão imediatamente.
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Standard'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Standard'
      security:
      - ApiKeyAuth: []
      summary: Revogar Token
      tags:
      - Auth
  /auth/logout-all:
    post:
      consumes:
      - application/json
      description: Incrementa a versão do token, invalidando todos os tokens do usuário.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Standard'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Standard'
      security:
      - ApiKeyAuth: []
      summary: Revogar todas as sessões
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Standard'
            - properties:
                data:
                  $ref: '#/definitions/auth.LoginResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Renovar tokens
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados para registro de novo usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Standard'
            - properties:
                data:
                  $ref: '#/definitions/auth.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Registra um novo usuário
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Recebe o token de reset e a nova senha para atualizar no DB.
      parameters:
      - description: Token de reset e nova senha
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Finalizar reset de senha
      tags:
      - Auth
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    description: Digite o token no formato Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"